Move the plots of the integration script into the integration test
